requirements.txt
# Motion-Triggered Recording Pipeline
opencv-python>=4.8.0
numpy>=1.24.0
python-dotenv>=1.0.0
PyYAML>=6.0

# Phase 1 dependencies (already installed)
pytapo>=3.3.18
nest_asyncio>=1.5.0
